Revert back to less aggressive component-level changes.
