+++ components/esmterrain/storage.cpp
@@ -16,6 +16,13 @@
 namespace ESMTerrain
 {
 
+    class LandCache
+    {
+    public:
+        typedef std::map<std::pair<int, int>, osg::ref_ptr<const LandObject> > Map;
+        Map mMap;
+    };
+
     LandObject::LandObject()
         : mLand(nullptr)
         , mLoadFlags(0)
@@ -89,6 +96,82 @@ namespace ESMTerrain
         min = defaultHeight;
         max = defaultHeight;
         return false;
+    }
+
+    void Storage::fixNormal (osg::Vec3f& normal, int cellX, int cellY, int col, int row, LandCache& cache)
+    {
+        while (col >= ESM::Land::LAND_SIZE-1)
+        {
+            ++cellY;
+            col -= ESM::Land::LAND_SIZE-1;
+        }
+        while (row >= ESM::Land::LAND_SIZE-1)
+        {
+            ++cellX;
+            row -= ESM::Land::LAND_SIZE-1;
+        }
+        while (col < 0)
+        {
+            --cellY;
+            col += ESM::Land::LAND_SIZE-1;
+        }
+        while (row < 0)
+        {
+            --cellX;
+            row += ESM::Land::LAND_SIZE-1;
+        }
+
+        const LandObject* land = getLand(cellX, cellY, cache);
+        const ESM::Land::LandData* data = land ? land->getData(ESM::Land::DATA_VNML) : 0;
+        if (data)
+        {
+            normal.x() = data->mNormals[col*ESM::Land::LAND_SIZE*3+row*3];
+            normal.y() = data->mNormals[col*ESM::Land::LAND_SIZE*3+row*3+1];
+            normal.z() = data->mNormals[col*ESM::Land::LAND_SIZE*3+row*3+2];
+            normal.normalize();
+        }
+        else
+            normal = osg::Vec3f(0,0,1);
+    }
+
+    void Storage::averageNormal(osg::Vec3f &normal, int cellX, int cellY, int col, int row, LandCache& cache)
+    {
+        osg::Vec3f n1,n2,n3,n4;
+        fixNormal(n1, cellX, cellY, col+1, row, cache);
+        fixNormal(n2, cellX, cellY, col-1, row, cache);
+        fixNormal(n3, cellX, cellY, col, row+1, cache);
+        fixNormal(n4, cellX, cellY, col, row-1, cache);
+        normal = (n1+n2+n3+n4);
+        normal.normalize();
+    }
+
+    void Storage::fixColour (osg::Vec4ub& color, int cellX, int cellY, int col, int row, LandCache& cache)
+    {
+        if (col == ESM::Land::LAND_SIZE-1)
+        {
+            ++cellY;
+            col = 0;
+        }
+        if (row == ESM::Land::LAND_SIZE-1)
+        {
+            ++cellX;
+            row = 0;
+        }
+
+        const LandObject* land = getLand(cellX, cellY, cache);
+        const ESM::Land::LandData* data = land ? land->getData(ESM::Land::DATA_VCLR) : 0;
+        if (data)
+        {
+            color.r() = data->mColours[col*ESM::Land::LAND_SIZE*3+row*3];
+            color.g() = data->mColours[col*ESM::Land::LAND_SIZE*3+row*3+1];
+            color.b() = data->mColours[col*ESM::Land::LAND_SIZE*3+row*3+2];
+        }
+        else
+        {
+            color.r() = 255;
+            color.g() = 255;
+            color.b() = 255;
+        }
     }
 
     void Storage::fillVertexBuffers (int lodLevel, float size, const osg::Vec2f& center,
@@ -172,7 +255,7 @@ namespace ESMTerrain
                         (*positions)[static_cast<unsigned int>(vertX*numVerts + vertY)]
                             = osg::Vec3f((vertX / float(numVerts - 1) - 0.5f) * size * Constants::CellSizeInUnits,
                                          (vertY / float(numVerts - 1) - 0.5f) * size * Constants::CellSizeInUnits,
-                                         height);
+                                         height + getAlteredHeight(col, row));
 
                         if (normalData)
                         {
@@ -207,6 +290,8 @@ namespace ESMTerrain
                             color.g() = 255;
                             color.b() = 255;
                         }
+
+                        adjustColor(col, row, heightData, color); //Does nothing by default, override in OpenMW-CS
 
                         // Unlike normals, colors mostly connect seamlessly between cells, but not always...
                         if (col == ESM::Land::LAND_SIZE-1 || row == ESM::Land::LAND_SIZE-1)
@@ -436,6 +521,27 @@ namespace ESMTerrain
                 -plane.getNormal().y() * nY
                 - plane[3]) / plane.getNormal().z() * Constants::CellSizeInUnits;
 
+    }
+
+    const LandObject* Storage::getLand(int cellX, int cellY, LandCache& cache)
+    {
+        LandCache::Map::iterator found = cache.mMap.find(std::make_pair(cellX, cellY));
+        if (found != cache.mMap.end())
+            return found->second;
+        else
+        {
+            found = cache.mMap.insert(std::make_pair(std::make_pair(cellX, cellY), getLand(cellX, cellY))).first;
+            return found->second;
+        }
+    }
+
+    void Storage::adjustColor(int col, int row, const ESM::Land::LandData *heightData, osg::Vec4ub& color) const
+    {
+    }
+
+    float Storage::getAlteredHeight(int col, int row) const
+    {
+        return 0;
     }
 
     Terrain::LayerInfo Storage::getLayerInfo(const std::string& texture)
